Move shared filters into the helper
